+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of Ffuenf_UpdateScripts.
 
+## 1.0.4 (February 3, 2016)
+
+* add helper for adding/updating cms pages, static blocks and email templates
+
 ## 1.0.3 (December 20, 2015)
 
 * fix dependency to Ffuenf_Common